Change: VisitMacrosが肥大化したのでカテゴリー毎に
